Move GeneratorType to `types.py`
