+++ jax_dataloader/types.py
@@ -1,0 +1,4 @@
+from .imports import *
+from .utils import Generator
+
+GeneratorType = Union[Generator, jax.Array, 'torch.Generator']
